requirements/development.txt
asgiref==3.4.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2.5
django-allauth==0.45.0
django-ckeditor==6.1.0
django-dotenv==1.4.2
django-email-verification==0.0.2
django-filter==2.4.0
django-js-asset==1.2.2
django-resized==0.3.11
django-tagging==0.5.0
idna==3.2
nodeenv==1.6.0
oauthlib==3.1.1
Pillow==8.3.1
pycparser==2.20
PyJWT==2.1.0
python-dotenv==0.19.0
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.26.6
django-appconf==1.0.4
django-compressor==2.4.1
django-sass-processor==1.0.1
libsass==0.21.0
rcssmin==1.0.6
rjsmin==1.1.0
django-rest-auth==0.9.5
PyJWT==1.7.1
djangorestframework-jwt==1.11.0
django-taggit==1.5.1
django-taggit-serializer==0.1.7

